refactor: Operation 제거
